invoker: wf-1782192502249-11/extract-sqlite-adapter-slices — Layer: persistence Feature state: active Goal: - Split `sqlite-adapter.ts` into schema/migrations, read queries, and write operations modules. Motivation: - Keep persistence logic maintainable and reduce risk in DB-affecting edits. Alternative considerations: - Leave persistence monolithic due to migration sensitivity. - Extract only query helpers and defer schema/write separation. Chosen design rationale: - Facade-preserving decomposition improves clarity without changing storage contract. Implementation details: - Keep `SQLiteAdapter` public surface stable while delegating internals to slice modules. Files: - packages/data-store/src/sqlite-adapter.ts - packages/data-store/src/sqlite-adapter/* Change types: - extract schema/migration helpers - extract read/write operation helpers Acceptance criteria: - Adapter public contract remains stable. - Migration and query behavior remains compatible.
Solution:
  Layer: persistence
Feature state: active
Goal:
- Split `sqlite-adapter.ts` into schema/migrations, read queries, and write operations modules.
Motivation:
- Keep persistence logic maintainable and reduce risk in DB-affecting edits.
Alternative considerations:
- Leave persistence monolithic due to migration sensitivity.
- Extract only query helpers and defer schema/write separation.
Chosen design rationale:
- Facade-preserving decomposition improves clarity without changing storage contract.
Implementation details:
- Keep `SQLiteAdapter` public surface stable while delegating internals to slice modules.
Files:
- packages/data-store/src/sqlite-adapter.ts
- packages/data-store/src/sqlite-adapter/*
Change types:
- extract schema/migration helpers
- extract read/write operation helpers
Acceptance criteria:
- Adapter public contract remains stable.
- Migration and query behavior remains compatible.
